--- Portafolio web/README.md
@@ -1,1 +1,0 @@
-Calla cholo de mrd
