+++ src/main/java/proyecto/web/trabajofinal/main.java
@@ -1,0 +1,13 @@
+package proyecto.web.trabajofinal;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class main {
+
+    public static void main(String[] args) {
+        SpringApplication.run(main.class, args);
+    }
+
+}
