+++ src/main/resources/templates/contacto.html
@@ -1,0 +1,182 @@
+<!DOCTYPE html>
+
+<html xmlns:th="http://www.thymeleaf.org">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" href="/style-contacto.css">
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Roboto:ital,wght@0,100..900;1,100..900&display=swap" rel="stylesheet">
+  <title>Andres Garza</title>
+</head>
+<body>
+<!-------WAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA----->
+<div class="container-color-botones">
+  <a class="boton-personalizacion-color" aria-expanded="false" aria-controls="menu-colores" title="Abrir selector de colores"><i class="fa-solid fa-pen" ></i></a>
+  <div class="menu-desplegable-colores" id="menu-colores" role="menu" aria-hidden="true" tabindex="-1">
+    <button class="circulo-color"  data-color="#e93535" role="listitem"></button>
+    <button class="circulo-color" data-color="#00cc00" role="listitem"></button>
+    <button class="circulo-color" data-color="#ff5e00" role="listitem"></button>
+    <button class="circulo-color" data-color="#23a7ff" role="listitem"></button>
+    <button class="circulo-color" data-color="#9333d3" role="listitem"></button>
+  </div>
+</div>
+<button id="color-toggle-btn" aria-label="Cambiar colores de fondo y texto" aria-pressed="false" type="button">
+  <i id="icon-sun" class="fa-solid fa-sun" style="display: inline-block;"></i>
+  <i id="icon-moon" class="fa-solid fa-moon" style="display: none;"></i>
+</button>
+<header>
+  <!--
+  Contiene el logo con enlace a la página principal, un botón para abrir el menú en dispositivos móviles
+  y una barra de navegación con enlaces a las secciones principales del portafolio: Inicio, Sobre mí, Proyectos y Contacto.
+  -->
+  <div class="container-header">
+    <a href="index.html" class="logo">Andres <span>Garza</span> </a>
+
+    <button class="btn-menu" aria-label="Abrir menú" aria-expanded="false"><i class="fa-solid fa-bars"></i></button>
+    <nav class="navbar">
+      <ul>
+        <li><a href="/index">Inicio</a></li>
+        <li><a href="/sobremi">Sobre mí</a></li>
+        <li><a href="/proyectos">Proyectos</a></li>
+        <li><a href="/contacto">Contacto</a></li>
+      </ul>
+    </nav>
+  </div>
+</header>
+<main>
+  <section class="contacto" id="contacto">
+    <div class="container-informacion">
+      <div class="informacion">
+        <h3 class="texto-ascendente-principal"> Trabajemos <span>juntos</span></h3>
+        <p class="texto-ascendente-principal">Prueba update Lorem ipsum, dolor sit amet consectetur adipisicing elit. Obcaecati fuga error quidem quis nemo deleniti dolores dolor reprehenderit autem beatae voluptatibus deserunt rem, repellendus sunt atque ullam adipisci ab expedita!</p>
+        <div class="bloques-info">
+          <div class="texto-info texto-ascendente-principal">
+            <div class="simbolo">
+              <i class="fa-solid fa-phone"></i></div>
+            <div class="noSimbolo">
+              <h4>Numero telefono</h4>
+              <p>+51 934 234 123</p>
+            </div>
+
+          </div>
+          <div class="texto-info texto-ascendente-principal">
+            <div class="simbolo">
+              <i class="fa-solid fa-location-dot"></i>
+            </div>
+
+            <div class="noSimbolo">
+              <h4>Officina</h4>
+              <p>calle 44, San Martin</p>
+            </div>
+          </div>
+          <div class="texto-info texto-ascendente-principal">
+            <div class="simbolo">
+              <i class="fa-solid fa-envelope"></i></i>
+            </div>
+            <div class="noSimbolo">
+              <h4>Correo</h4>
+              <p>micorreo@gmail.com</p>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="container-formulario">
+      <div class="fondo-formulario">
+        <h3 class="texto-ascendente-sobremi">Enviame un <span>correo</span></h3>
+        <form id="form" action="">
+          <div class="input-grupo">
+            <div class="input-caja">
+              <input type="text" placeholder="Nombre Completo" name="nombre" class="texto-ascendente-sobremi" required>
+              <input type="email" placeholder="Correo Electronico" name="correo-electronico" class="texto-ascendente-sobremi" required>
+            </div>
+            <div class="input-caja">
+              <input type="number" placeholder="Numero de celular" name="celular" class="texto-ascendente-sobremi"  required>
+              <input type="text" placeholder="Asunto" name="asunto" class="texto-ascendente-sobremi"  required>
+            </div>
+          </div>
+          <div class="input-grupo-2">
+            <textarea name="" id="" cols="30" rows="10" placeholder="Escribe tu mensaje aqui>" class="texto-ascendente-sobremi"  required></textarea>
+            <div class="texto-ascendente-sobremi">
+              <input id="button-form" type="submit" value="Send Message" class="btn-general btn-color btn-enviar ">
+            </div>
+
+          </div>
+        </form>
+
+      </div>
+
+    </div>
+  </section>
+
+</main>
+<footer>
+  <div class="container-footer">
+    <div class="footer-superior">
+      <div class="container-mapa">
+        <h3>Mi <span>ubicación</span></h3>
+        <div id="map"></div>
+      </div>
+      <div class="texto-footer">
+        <div class="footer-menu">
+          <h4>Menu</h4>
+          <ul class="lista">
+            <li>
+              <a href="#">Index</a>
+            </li>
+            <li>
+              <a href="#sobremi">Sobre mi</a>
+            </li>
+            <li>
+              <a href="#Proyectos">Proyectos</a>
+            </li>
+            <li>
+              <a href="#contacto"> contacto</a>
+            </li>
+          </ul>
+        </div>
+
+        <div class="footer-contacto">
+          <h4>Contacto</h4>
+          <div class="footer-info">
+            <i class="fa-solid fa-phone"></i>
+            <p>+51 934 234 123</p>
+          </div>
+          <div class="footer-info">
+            <i class="fa-solid fa-location-dot"></i>
+            <p>calle 44, San Martin</p>
+          </div>
+          <div class="footer-info">
+            <i class="fa-solid fa-envelope"></i></i>
+            <p>micorreo@gmail.com</p>
+          </div>
+        </div>
+        <div class="redes redes-footer">
+          <h4>redes</h4>
+          <a href=""><i class="fab fa-facebook-f"></i></a>
+          <a href="#"><i class="fab fa-linkedin"></i></a>
+          <a href="#"><i class="fab fa-github"></i></a>
+          <a href="#"><i class="fab fa-twitter"></i></a>
+          <a href="#"><i class="fab fa-instagram"></i></a>
+        </div>
+
+
+      </div>
+    </div>
+    <div class="footer-inferior">
+      <p class="copyright">
+        © Andres Garza | Todos los derechos reservados.
+      </p>
+    </div>
+  </div>
+</footer>
+<script type="text/javascript" src="https://cdn.jsdelivr.net/npm/@emailjs/browser@4/dist/email.min.js"></script>
+<script type="text/javascript">emailjs.init('N-Qj4Yxgsu8nK4poX')</script>
+<script src="/script.js"></script>
+<script async defer src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBDpLiSpzD3YXJeZf7flL7cQnBCKyygQSI&callback=initMap"></script>
+</body>
+</html>
