+++ src/main/resources/templates/editor.html
@@ -1,0 +1,90 @@
+<!DOCTYPE html>
+<html lang="es">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <link rel="stylesheet" href="style-editor.css">
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Roboto:ital,wght@0,100..900;1,100..900&display=swap" rel="stylesheet">
+  <title>Andres Garza</title>
+</head>
+<body>
+<div class="container-color-botones">
+  <a class="boton-personalizacion-color" aria-expanded="false" aria-controls="menu-colores" title="Abrir selector de colores"><i class="fa-solid fa-pen" ></i></a>
+  <div class="menu-desplegable-colores" id="menu-colores" role="menu" aria-hidden="true" tabindex="-1">
+    <button class="circulo-color"  data-color="#e93535" role="listitem"></button>
+    <button class="circulo-color" data-color="#00cc00" role="listitem"></button>
+    <button class="circulo-color" data-color="#ff5e00" role="listitem"></button>
+    <button class="circulo-color" data-color="#23a7ff" role="listitem"></button>
+    <button class="circulo-color" data-color="#9900ff" role="listitem"></button>
+  </div>
+</div>
+<button id="color-toggle-btn" aria-label="Cambiar colores de fondo y texto" aria-pressed="false" type="button">
+  <i id="icon-sun" class="fa-solid fa-sun" style="display: inline-block;"></i>
+  <i id="icon-moon" class="fa-solid fa-moon" style="display: none;"></i>
+</button>
+
+<main>
+  <section class="form-container" aria-labelledby="form-title">
+    <h2 id="form-title" class="form-title">Cambiar datos</h2>
+    <form id="data-form" class="data-form" novalidate>
+      <div class="image-upload-wrapper" aria-label="Subir imagen de perfil">
+        <input type="file" id="image-upload" accept="image/*" hidden />
+        <label
+                for="image-upload"
+                class="image-upload-circle"
+                tabindex="0"
+                aria-describedby="upload-desc"
+                role="button"
+        >
+          <div class="image-preview-container">
+            <img
+                    src=""
+                    alt="Previsualización de la imagen subida"
+                    class="image-preview"
+                    aria-hidden="true"
+            />
+          </div>
+          <span id="upload-desc" class="visually-hidden"
+          >Click o arrastra imagen aquí para subir</span
+          >
+          <div class="image-action-buttons" aria-hidden="true">
+            <button
+                    type="button"
+                    class="btn-icon btn-remove"
+                    title="Eliminar imagen"
+                    aria-label="Eliminar imagen"
+            >
+              <i class="fas fa-times"></i>
+            </button>
+            <button
+                    type="button"
+                    class="btn-icon btn-rotate"
+                    title="Girar imagen"
+                    aria-label="Girar imagen"
+            >
+              <i class="fas fa-rotate-right"></i>
+            </button>
+          </div>
+          <div class="upload-placeholder" aria-hidden="true">
+            <i class="fas fa-cloud-upload-alt"></i>
+          </div>
+        </label>
+      </div>
+      <div id="form-fields-container">
+        <!-- Aquí se agregarán dinámicamente los nuevos campos verticalmente -->
+      </div>
+      <button type="submit" class="btn-save">Guardar</button>
+    </form>
+  </section>
+</main>
+
+
+
+<a href="index.html"> Volver a la pagina principal</a>
+<script src="script.js"></script>
+</body>
+</html>
