+++ src/test/java/proyecto/web/trabajofinal/TrabajoFinalApplicationTests.java
@@ -1,0 +1,13 @@
+package proyecto.web.trabajofinal;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TrabajoFinalApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
